requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
django-redis==6.0.0
django-rest-framework==0.1.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
uritemplate==4.2.0
